权限调整

+++ src/main/java/com/crazy/portal/dao/system/ResourceMapper.java
@@ -42,5 +42,5 @@ public interface ResourceMapper {
      */
     int countOrder();
 
-    int getRoleCountByResourceId(Long resourceId);
+    int getRoleCountByResourceId(Integer resourceId);
 }
